requirements.txt
django~=4.2.0
django-environ==0.11.2
django-bootstrap5==24.2
djangorestframework==3.15.1
markdown==3.6
django-filter==24.2
django-taggit==5.0.1

tqdm
pillow == 10.3.0

ipython
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-cors-headers==4.3.1